category edit dialog updates + overall formating and minor file reorg
